@@ -4,7 +4,7 @@ version: 1.0.3
 author: Realank <realank@126.com>
 homepage: https://github.com/Realank/flutter_section_table_view
 environment:
-    sdk: ">=2.0.0-dev.28.0 <3.0.0"
+    sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
