@@ -16,8 +16,8 @@ typedef double CellHeightAtIndexPathCallBack(int section, int row);
 typedef void SectionTableViewScrollToCallBack(int section, int row, bool isScrollDown);
 
 class IndexPath {
-  final int section;
-  final int row;
+  final int? section;
+  final int? row;
   IndexPath({this.section, this.row});
   @override
   String toString() {
@@ -25,22 +25,22 @@ class IndexPath {
   }
 
   @override
-  int get hashCode => super.hashCode;
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is IndexPath &&
+          runtimeType == other.runtimeType &&
+          section == other.section &&
+          row == other.row;
+
   @override
-  bool operator ==(other) {
-    if (other.runtimeType != IndexPath) {
-      return false;
-    }
-    IndexPath otherIndexPath = other;
-    return section == otherIndexPath.section && row == otherIndexPath.row;
-  }
+  int get hashCode => section.hashCode ^ row.hashCode;
 }
 
 class SectionTableController extends ChangeNotifier {
   IndexPath topIndex = IndexPath(section: 0, row: -1);
   bool dirty = false;
   bool animate = false;
-  SectionTableViewScrollToCallBack sectionTableViewScrollTo;
+  SectionTableViewScrollToCallBack? sectionTableViewScrollTo;
 
   SectionTableController({this.sectionTableViewScrollTo});
 
@@ -67,16 +67,16 @@ class SectionTableView extends StatefulWidget {
   final CellAtIndexPathCallBack cellAtIndexPath;
 
   //section header & divider
-  final SectionHeaderCallBack headerInSection;
-  final Widget divider;
+  final SectionHeaderCallBack? headerInSection;
+  final Widget? divider;
 
   //tell me cell & header & divider height, so that I can scroll to specific index path
   //work with SectionTableController
-  final SectionHeaderHeightCallBack sectionHeaderHeight; // must set when use SectionTableController
-  final DividerHeightCallBack dividerHeight; // must set when use SectionTableController
-  final CellHeightAtIndexPathCallBack
+  final SectionHeaderHeightCallBack? sectionHeaderHeight; // must set when use SectionTableController
+  final DividerHeightCallBack? dividerHeight; // must set when use SectionTableController
+  final CellHeightAtIndexPathCallBack?
       cellHeightAtIndexPath; // must set when use SectionTableController
-  final SectionTableController
+  final SectionTableController?
       controller; //you can use this controller to scroll section table view
 
   //pull refresh
@@ -87,17 +87,17 @@ class SectionTableView extends StatefulWidget {
   final Config refreshHeaderConfig, refreshFooterConfig; // configure your refresh header and footer
   final bool enablePullUp;
   final bool enablePullDown;
-  final OnRefresh onRefresh;
+  final OnRefresh? onRefresh;
 
-  final ScrollController _scrollController;
-  final RefreshController refreshController;
-  ScrollController get scrollController => _scrollController;
+  final ScrollController? _scrollController;
+  final RefreshController? refreshController;
+  ScrollController get scrollController => _scrollController!;
 
   SectionTableView({
-    Key key,
-    @required this.sectionCount,
-    @required this.numOfRowInSection,
-    @required this.cellAtIndexPath,
+    Key? key,
+    required this.sectionCount,
+    required this.numOfRowInSection,
+    required this.cellAtIndexPath,
     this.headerInSection,
     this.divider,
     this.sectionHeaderHeight,
@@ -131,7 +131,7 @@ class SectionTableView extends StatefulWidget {
             }),
         assert((enablePullDown || enablePullUp) ? refreshController != null : true),
         _scrollController = (enablePullDown || enablePullUp)
-            ? refreshController.scrollController
+            ? refreshController?.scrollController
             : ScrollController(),
         super(key: key);
   @override
@@ -140,18 +140,18 @@ class SectionTableView extends StatefulWidget {
 
 class _SectionTableViewState extends State<SectionTableView> {
   List<IndexPath> indexToIndexPathSearch = [];
-  Map<String, double> indexPathToOffsetSearch;
+  Map<String, double> indexPathToOffsetSearch = {};
 
   final listViewKey = GlobalKey();
 
   //scroll position check
-  int currentIndex;
-  double preIndexOffset;
-  double nextIndexOffset;
+  int? currentIndex;
+  double? preIndexOffset;
+  double? nextIndexOffset;
 
-  bool showDivider;
+  bool showDivider = false;
 
-  double scrollOffsetFromIndex(IndexPath indexPath) {
+  double? scrollOffsetFromIndex(IndexPath indexPath) {
     var offset = indexPathToOffsetSearch[indexPath.toString()];
     if (offset == null) {
       return null;
@@ -160,7 +160,7 @@ class _SectionTableViewState extends State<SectionTableView> {
         indexPathToOffsetSearch[IndexPath(section: widget.sectionCount, row: -1).toString()];
 
     if (listViewKey.currentContext != null && contentHeight != null) {
-      double listViewHeight = listViewKey.currentContext.size.height;
+      double listViewHeight = listViewKey.currentContext?.size?.height??0;
       if (widget.enablePullUp) {
         listViewHeight -= 60.0; //refresh header height
       }
@@ -218,16 +218,16 @@ class _SectionTableViewState extends State<SectionTableView> {
                [sectionHeaderHeight][dividerHeight][cellHeightAtIndexPath]''');
     } else {
       double offset = 0.0;
-      double dividerHeight = showDivider ? widget.dividerHeight() : 0.0;
+      double dividerHeight = showDivider ? (widget.dividerHeight!()) : 0.0;
       for (int i = 0; i < widget.sectionCount; i++) {
         if (showSectionHeader) {
           indexPathToOffsetSearch[IndexPath(section: i, row: -1).toString()] = offset;
-          offset += widget.sectionHeaderHeight(i);
+          offset += widget.sectionHeaderHeight!(i);
         }
         int rows = widget.numOfRowInSection(i);
         for (int j = 0; j < rows; j++) {
           indexPathToOffsetSearch[IndexPath(section: i, row: j).toString()] = offset;
-          offset += widget.cellHeightAtIndexPath(i, j) + dividerHeight;
+          offset += widget.cellHeightAtIndexPath!(i, j) + dividerHeight;
         }
       }
       indexPathToOffsetSearch[IndexPath(section: widget.sectionCount, row: -1).toString()] =
@@ -243,7 +243,7 @@ class _SectionTableViewState extends State<SectionTableView> {
     int findValidIndexPathByIndex(int index, int pace) {
       for (int i = index + pace; (i >= 0 && i < indexToIndexPathSearch.length); i += pace) {
         final indexPath = indexToIndexPathSearch[i];
-        if (indexPath.section >= 0) {
+        if ((indexPath.section??0) >= 0) {
           return i;
         }
       }
@@ -257,28 +257,28 @@ class _SectionTableViewState extends State<SectionTableView> {
     if (indexPathToOffsetSearch != null) {
       currentIndex = 0;
       for (int i = 0; i < indexToIndexPathSearch.length; i++) {
-        if (indexToIndexPathSearch[i] == sectionController.topIndex) {
+        if (indexToIndexPathSearch[i] == sectionController?.topIndex) {
           currentIndex = i;
         }
       }
 
 //      final preIndexPath = findValidIndexPathByIndex(currentIndex, -1);
-      final currentIndexPath = indexToIndexPathSearch[currentIndex];
-      final nextIndexPath = indexToIndexPathSearch[findValidIndexPathByIndex(currentIndex, 1)];
+      final currentIndexPath = indexToIndexPathSearch[currentIndex!];
+      final nextIndexPath = indexToIndexPathSearch[findValidIndexPathByIndex(currentIndex!, 1)];
       preIndexOffset = indexPathToOffsetSearch[currentIndexPath.toString()];
       nextIndexOffset = indexPathToOffsetSearch[nextIndexPath.toString()];
     }
 
     //init scroll controller
-    widget.controller.addListener(() {
+    widget.controller?.addListener(() {
       //listen section table controller to scroll the list view
-      if (sectionController.dirty) {
-        sectionController.dirty = false;
-        double offset = scrollOffsetFromIndex(sectionController.topIndex);
+      if (sectionController?.dirty==true) {
+        sectionController?.dirty = false;
+        double? offset = scrollOffsetFromIndex(sectionController!.topIndex);
         if (offset == null) {
           return;
         }
-        if (sectionController.animate) {
+        if (sectionController?.animate==true) {
           widget.scrollController
               .animateTo(offset, duration: Duration(milliseconds: 250), curve: Curves.decelerate);
         } else {
@@ -291,33 +291,32 @@ class _SectionTableViewState extends State<SectionTableView> {
       widget.scrollController.addListener(() {
         double currentOffset = widget.scrollController.offset;
 //        print('scroll offset $currentOffset');
-        if (currentOffset < preIndexOffset) {
+        if (currentOffset < (preIndexOffset??0)) {
           //go previous cell
-          if (currentIndex > 0) {
-            final nextIndexPath = indexToIndexPathSearch[currentIndex];
-            currentIndex = findValidIndexPathByIndex(currentIndex, -1);
-            final currentIndexPath = indexToIndexPathSearch[currentIndex];
+          if ((currentIndex??0) > 0) {
+            final nextIndexPath = indexToIndexPathSearch[currentIndex!];
+            currentIndex = findValidIndexPathByIndex(currentIndex!, -1);
+            final currentIndexPath = indexToIndexPathSearch[currentIndex!];
             preIndexOffset = indexPathToOffsetSearch[currentIndexPath.toString()];
             nextIndexOffset = indexPathToOffsetSearch[nextIndexPath.toString()];
 //            print('go previous index $currentIndexPath');
-            if (widget.controller.sectionTableViewScrollTo != null) {
-              widget.controller
-                  .sectionTableViewScrollTo(currentIndexPath.section, currentIndexPath.row, false);
+            if (widget.controller?.sectionTableViewScrollTo != null) {
+              widget.controller?.sectionTableViewScrollTo!(currentIndexPath.section??0, currentIndexPath.row??0, false);
             }
           }
-        } else if (currentOffset >= nextIndexOffset) {
+        } else if (currentOffset >= (nextIndexOffset??0)) {
           //go next cell
-          if (currentIndex < indexToIndexPathSearch.length - 2) {
-            currentIndex = findValidIndexPathByIndex(currentIndex, 1);
-            final currentIndexPath = indexToIndexPathSearch[currentIndex];
+          if ((currentIndex??0) < indexToIndexPathSearch.length - 2) {
+            currentIndex = findValidIndexPathByIndex(currentIndex!, 1);
+            final currentIndexPath = indexToIndexPathSearch[currentIndex!];
             final nextIndexPath =
-                indexToIndexPathSearch[findValidIndexPathByIndex(currentIndex, 1)];
+                indexToIndexPathSearch[findValidIndexPathByIndex(currentIndex!, 1)];
             preIndexOffset = indexPathToOffsetSearch[currentIndexPath.toString()];
             nextIndexOffset = indexPathToOffsetSearch[nextIndexPath.toString()];
 //            print('go next index $currentIndexPath');
-            if (widget.controller.sectionTableViewScrollTo != null) {
-              widget.controller
-                  .sectionTableViewScrollTo(currentIndexPath.section, currentIndexPath.row, true);
+            if (widget.controller?.sectionTableViewScrollTo != null) {
+              widget.controller!
+                  .sectionTableViewScrollTo!(currentIndexPath.section!, currentIndexPath.row!, true);
             }
           }
         }
@@ -354,14 +353,14 @@ class _SectionTableViewState extends State<SectionTableView> {
 
     IndexPath indexPath = indexToIndexPathSearch[index];
     //section header
-    if (indexPath.section >= 0 && indexPath.row < 0) {
-      return widget.headerInSection(indexPath.section);
+    if ((indexPath.section??0) >= 0 && (indexPath.row??0) < 0) {
+      return widget.headerInSection!(indexPath.section!);
     }
 
-    Widget cell = widget.cellAtIndexPath(indexPath.section, indexPath.row);
+    Widget cell = widget.cellAtIndexPath(indexPath.section!, indexPath.row!);
     if (showDivider) {
       return Column(
-        children: <Widget>[cell, widget.divider],
+        children: <Widget>[cell, widget.divider??Container()],
         mainAxisSize: MainAxisSize.min,
       );
     } else {
