@@ -34,8 +34,8 @@ class ClassicIndicator extends Indicator {
   final TextStyle textStyle;
 
   const ClassicIndicator({
-    @required int mode,
-    Key key,
+    required int mode,
+    Key? key,
     this.textStyle: const TextStyle(color: const Color(0xff555555)),
     this.releaseText: 'Refresh when release',
     this.refreshingText: 'Refreshing...',
