@@ -45,18 +45,18 @@ class SmartRefresher extends StatefulWidget {
   // if open OverScroll if you use RefreshIndicator and LoadFooter
   final bool enableOverScroll;
   // upper and downer callback when you drag out of the distance
-  final OnRefresh onRefresh;
+  final OnRefresh? onRefresh;
   // This method will callback when the indicator changes from edge to edge.
-  final OnOffsetChange onOffsetChange;
+  final OnOffsetChange? onOffsetChange;
   //controll inner state
   final RefreshController controller;
 
   SmartRefresher({
-    Key key,
-    @required this.child,
-    IndicatorBuilder headerBuilder,
-    IndicatorBuilder footerBuilder,
-    RefreshController controller,
+    Key? key,
+    required this.child,
+    IndicatorBuilder? headerBuilder,
+    IndicatorBuilder? footerBuilder,
+    RefreshController? controller,
     this.headerConfig: const RefreshConfig(),
     this.footerConfig: const LoadConfig(),
     this.enableOverScroll: default_enableOverScroll,
@@ -165,13 +165,13 @@ class _SmartRefresherState extends State<SmartRefresher> {
         0) {
       return (notification.metrics.minScrollExtent -
               notification.metrics.pixels) /
-          widget.headerConfig.triggerDistance;
+          (widget.headerConfig.triggerDistance??0);
     } else if (notification.metrics.pixels -
             notification.metrics.maxScrollExtent >
         0) {
       return (notification.metrics.pixels -
               notification.metrics.maxScrollExtent) /
-          widget.footerConfig.triggerDistance;
+          (widget.footerConfig.triggerDistance??0);
     }
     return -1.0;
   }
@@ -209,21 +209,21 @@ class _SmartRefresherState extends State<SmartRefresher> {
     if (overscrollPastStart > overscrollPastEnd) {
       if (widget.headerConfig is RefreshConfig) {
         if (widget.onOffsetChange != null) {
-          widget.onOffsetChange(true, overscrollPastStart);
+          widget.onOffsetChange!(true, overscrollPastStart);
         }
       } else {
         if (widget.onOffsetChange != null) {
-          widget.onOffsetChange(true, overscrollPastStart);
+          widget.onOffsetChange!(true, overscrollPastStart);
         }
       }
     } else if (overscrollPastEnd > 0) {
       if (widget.footerConfig is RefreshConfig) {
         if (widget.onOffsetChange != null) {
-          widget.onOffsetChange(false, overscrollPastEnd);
+          widget.onOffsetChange!(false, overscrollPastEnd);
         }
       } else {
         if (widget.onOffsetChange != null) {
-          widget.onOffsetChange(false, overscrollPastEnd);
+          widget.onOffsetChange!(false, overscrollPastEnd);
         }
       }
     }
@@ -233,7 +233,7 @@ class _SmartRefresherState extends State<SmartRefresher> {
     switch (mode.value) {
       case RefreshStatus.refreshing:
         if (widget.onRefresh != null) {
-          widget.onRefresh(up);
+          widget.onRefresh!(up);
         }
         if (up && widget.headerConfig is RefreshConfig) {
           RefreshConfig config = widget.headerConfig as RefreshConfig;
@@ -260,9 +260,9 @@ class _SmartRefresherState extends State<SmartRefresher> {
     });
     setState(() {
       if (widget.enablePullDown)
-        _headerHeight = _headerKey.currentContext.size.height;
+        _headerHeight = _headerKey.currentContext?.size?.height??0;
       if (widget.enablePullUp) {
-        _footerHeight = _footerKey.currentContext.size.height;
+        _footerHeight = _footerKey.currentContext?.size?.height??0;
       }
     });
   }
@@ -297,7 +297,7 @@ class _SmartRefresherState extends State<SmartRefresher> {
         up: up,
         onOffsetChange: (bool up, double offset) {
           if (widget.onOffsetChange != null) {
-            widget.onOffsetChange(
+            widget.onOffsetChange!(
                 up,
                 up
                     ? -_scrollController.offset + offset
@@ -367,15 +367,15 @@ class _SmartRefresherState extends State<SmartRefresher> {
 }
 
 abstract class Indicator extends StatefulWidget {
-  final int mode;
+  final int? mode;
 
-  const Indicator({Key key, this.mode}) : super(key: key);
+  const Indicator({Key? key, this.mode}) : super(key: key);
 }
 
 class RefreshController {
-  ValueNotifier<int> _headerMode;
-  ValueNotifier<int> _footerMode;
-  ScrollController scrollController;
+  ValueNotifier<int>? _headerMode;
+  ValueNotifier<int>? _footerMode;
+  late ScrollController scrollController;
 
   RefreshController() {
     scrollController = ScrollController();
@@ -383,11 +383,11 @@ class RefreshController {
 
   void requestRefresh(bool up) {
     if (up) {
-      if (_headerMode.value == RefreshStatus.idle)
-        _headerMode.value = RefreshStatus.refreshing;
+      if (_headerMode?.value == RefreshStatus.idle)
+        _headerMode?.value = RefreshStatus.refreshing;
     } else {
-      if (_footerMode.value == RefreshStatus.idle) {
-        _footerMode.value = RefreshStatus.refreshing;
+      if (_footerMode?.value == RefreshStatus.idle) {
+        _footerMode?.value = RefreshStatus.refreshing;
       }
     }
   }
@@ -396,31 +396,31 @@ class RefreshController {
     scrollController.jumpTo(offset);
   }
 
-  Future<Null> animateTo(
+  Future<void> animateTo(
     double offset, {
-    @required Duration duration,
-    @required Curve curve,
+    required Duration duration,
+    required Curve curve,
   }) {
     return scrollController.animateTo(offset, duration: duration, curve: curve);
   }
 
   void sendBack(bool up, int mode) {
     if (up) {
-      _headerMode.value = mode;
+      _headerMode?.value = mode;
     } else {
-      _footerMode.value = mode;
+      _footerMode?.value = mode;
     }
   }
 
-  int get headerMode => _headerMode.value;
+  int get headerMode => _headerMode?.value??0;
 
-  int get footerMode => _footerMode.value;
+  int get footerMode => _footerMode?.value??0;
 
   isRefresh(bool up) {
     if (up) {
-      return _headerMode.value == RefreshStatus.refreshing;
+      return _headerMode?.value == RefreshStatus.refreshing;
     } else {
-      return _footerMode.value == RefreshStatus.refreshing;
+      return _footerMode?.value == RefreshStatus.refreshing;
     }
   }
 }
